Isolated search params, fix for dive datetime  issue, fix computerid issue
