+++ src/Components/DiveSearch/types.ts
@@ -1,11 +1,4 @@
-export enum SearchOptions {
-    Keyword = 'keyword',
-    Tag = 'tag',
-    Buddy = 'buddy',
-    Before = 'before',
-    After = 'after',
-    Place = 'place',
-}
+import { SearchOptions } from './searchOptions';
 
 export type SearchFilter = { type: SearchOptions; value: string | number };
 export type SearchFilters = SearchFilter[];
